CONNECTION OKAY TO WS GAME:
