+++ front/app/src/main.tsx
@@ -8,7 +8,7 @@ import { Provider } from "react-redux";
 import { PersistGate } from "redux-persist/integration/react";
 
 // TEST
-localStorage.clear();
+// localStorage.clear();
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
 		<Provider store={store}>
